Updating the mitosis classification code.
This fixes a major bug affecting resuming of training in which the base
Keras model would not be marked as initialized, and therefore calls to
`model.save` would call `K.get_session()`, which would reinitialize
those variables.  This also cleans up the experiment names, and a few
other minor improvements.